main_consumer/requirements.txt
confluent_kafka==2.3.0
Django==4.2.3
django_elasticsearch_dsl==8.0
djangorestframework==3.14.0
elasticsearch==8.12.1
elasticsearch_dsl==8.12.0
Requests==2.31.0
